fix problem with manually loading columns
