+++ app/__tests__/parse-settings.test.ts
@@ -13,6 +13,10 @@ import {
 import * as parseSettings from "../src/state/parse-settings";
 import { ChartSettings } from "../src/types";
 import { readFileSync } from "fs";
+vi.mock("../src/state/actions/util-actions", () => ({
+  changeManuallyAddedColumns: vi.fn(),
+  normalizeColumnProperties: vi.fn()
+}));
 const tests = JSON.parse(readFileSync("./app/__tests__/__data__/parse-settings-tests.json").toString());
 const keys = JSON.parse(readFileSync("./app/__tests__/__data__/parse-settings-keys.json").toString());
 describe("escape html chars", () => {
@@ -603,6 +607,6 @@ describe("json to xml", () => {
       `    <column id="class datastore.RootColumn:Chart Root">\n` +
       `    </column>\n` +
       `</TSCreator>`;
-    expect(parseSettings.jsonToXml({} as ColumnInfo, {} as ChartSettings)).toEqual(key);
+    expect(parseSettings.jsonToXml({} as ColumnInfo, new Map<string, ColumnInfo>(), {} as ChartSettings)).toEqual(key);
   });
 });
